separated api and network class
