+++ src/qgis_stac/api/base.py
@@ -1,7 +1,4 @@
-import dataclasses
 import typing
-
-from functools import partial
 
 from qgis.PyQt import (
     QtCore,
@@ -9,15 +6,10 @@ from qgis.PyQt import (
     QtXml
 )
 
-from qgis.core import (
-    QgsApplication,
-    QgsNetworkContentFetcherTask,
-    QgsTask,
-)
-
 import models
 
-from qgis_stac.lib.pystac_client import Client
+from network import NetworkFetchTask
+from qgis_stac.conf import ConnectionSettings
 
 
 class BaseClient(QtCore.QObject):
@@ -25,8 +17,14 @@ class BaseClient(QtCore.QObject):
     url: str
     network_task: NetworkFetchTask
 
-    collections_received = QtCore.pyqtSignal(list, models.ResourcePagination)
-    items_received = QtCore.pyqtSignal(list, models.ResourcePagination)
+    collections_received = QtCore.pyqtSignal(
+        list[models.Collection],
+        models.ResourcePagination
+    )
+    items_received = QtCore.pyqtSignal(
+        list[models.Item],
+        models.ResourcePagination
+    )
     error_received = QtCore.pyqtSignal([str], [str, int, str])
 
     def __init__(
@@ -44,7 +42,7 @@ class BaseClient(QtCore.QObject):
     @classmethod
     def from_connection_settings(
             cls,
-            connection_settings: "ConnectionSettings"
+            connection_settings: ConnectionSettings
     ):
         return cls(
             url=connection_settings.url,
@@ -55,19 +53,30 @@ class BaseClient(QtCore.QObject):
         self,
         item_search: typing.Optional[models.ItemSearch] = None
     ):
-        url = self.get_item_search_url(item_search)
-        request = QtNetwork.QNetworkRequest(url)
+        params = self.get_search_params(item_search)
         self.network_task = NetworkFetchTask(
             url=self.url,
-            search_params=item_search,
+            search_params=params,
             resource_type=models.ResourceType.FEATURE,
-            response_handler=self.handle_items(),
-            error_handler=self.handle_error(),
+            response_handler=self.handle_items,
+            error_handler=self.handle_error,
         )
+
+    def get_items_search_params(self, params: models.ItemSearch):
+        search_params = {
+            ''
+        }
+
+
+    def handle_collections(
+            self,
+            collections
+    ):
+        raise NotImplementedError
 
     def handle_items(
             self,
-            payload: typing.Any
+            items
     ):
         raise NotImplementedError
 
@@ -76,53 +85,3 @@ class BaseClient(QtCore.QObject):
             message: str
     ):
         raise NotImplementedError
-
-
-class NetworkFetchTask(QgsTask):
-
-    url: str
-    search_params: models.ItemSearch
-    resource_type: models.ResourceType
-    response_handler: typing.Callable
-    error_handler: typing.Callable
-
-    response: QtCore.QByteArray = None
-    client: Client = None
-
-    def __init__(
-        self,
-        url: str,
-        search_params: models.ItemSearch,
-        resource_type: models.ResourceType,
-        response_handler: typing.Callable = None,
-        error_handler: typing.Callable = None,
-    ):
-        """
-        """
-        super().__init__()
-        self.url = url
-        self.search_params = search_params
-        self.resource_type = resource_type
-        self.response_handler = response_handler
-        self.error_handler = error_handler
-
-    def run(self):
-        self.client = Client.open(self.url)
-        if self.resource_type ==\
-                models.ResourceType.FEATURE:
-            self.response = self.client.search(
-                self.search_params.to_dict())
-        elif self.resource_type == \
-                models.ResourceType.COLLECTION:
-            self.response = self.client.get_collections()
-        else:
-            raise NotImplementedError
-
-        return self.response is not None
-
-    def finished(self, result: bool):
-        if result:
-            self.response_handler(self.response)
-        else:
-            message = f"Error fetching content for {self.url!r}"
-            self.error_handler(message)
